Adding -R option to plough, for repeatedly running the same experiment

+++ tractor.utils/R/plough.R
@@ -1,15 +1,22 @@
-ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, tractorOptions, useGridEngine, crossApply, queueName, qsubOptions, parallelisationFactor, debug)
+ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, tractorOptions, useGridEngine, crossApply, queueName, qsubOptions, parallelisationFactor, debug, repetitions)
 {
     crossApply <- isTRUE(crossApply == 1)
     useGridEngine <- isTRUE(useGridEngine == 1)
     setOutputLevel(ifelse(isTRUE(debug==1), OL$Debug, OL$Info))
+    
+    if (isValidAs(repetitions, "integer"))
+        repetitions <- as.integer(repetitions)
+    else
+        repetitions <- 0L
     
     config <- readYaml(configFiles)
     variableLengths <- sapply(config, length)
     
     variables <- splitAndConvertString(variables, ",", fixed=TRUE)
     variables <- variables[variables != ""]
-    if (length(variables) == 0)
+    if (repetitions > 0)
+        variables <- character(0)
+    else if (length(variables) == 0)
     {
         if (any(variableLengths > 1))
             variables <- names(config)[variableLengths > 1]
@@ -19,7 +26,8 @@ ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, 
     else if (!all(variables %in% names(config)))
         report(OL$Error, "Specified variable(s) #{implode(variables[!(variables %in% names(config))],sep=', ',finalSep=' and ')} are not mentioned in the config files")
     
-    report(OL$Info, "Looping over variable(s) #{implode(variables,sep=', ',finalSep=' and ')}")
+    if (length(variables) > 0)
+        report(OL$Info, "Looping over variable(s) #{implode(variables,sep=', ',finalSep=' and ')}")
     
     usingParallel <- FALSE
     if (isValidAs(parallelisationFactor,"integer") && as.integer(parallelisationFactor) > 1)
@@ -36,19 +44,30 @@ ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, 
             report(OL$Warning, "The \"parallel\" package is not installed - code will not be parallelised")
     }
     
-    if (crossApply)
+    data <- configFile <- NULL
+    if (repetitions > 0)
     {
-        n <- prod(variableLengths[variables])
-        data <- as.data.frame(expand.grid(config[variables], stringsAsFactors=FALSE), stringsAsFactors=FALSE)
+        n <- as.integer(repetitions)
+        data <- as.data.frame(config, stringsAsFactors=FALSE)
+        configFile <- ensureFileSuffix(threadSafeTempFile(), "yaml")
+        writeYaml(config, configFile)
     }
     else
     {
-        n <- max(variableLengths[variables])
-        data <- as.data.frame(lapply(config[variables], rep, length.out=n), stringsAsFactors=FALSE)
-    }
+        if (crossApply)
+        {
+            n <- prod(variableLengths[variables])
+            data <- as.data.frame(expand.grid(config[variables], stringsAsFactors=FALSE), stringsAsFactors=FALSE)
+        }
+        else
+        {
+            n <- max(variableLengths[variables])
+            data <- as.data.frame(lapply(config[variables], rep, length.out=n), stringsAsFactors=FALSE)
+        }
     
-    for (name in setdiff(names(config), c(variables,"")))
-        data[[name]] <- rep(config[[name]], length.out=n)
+        for (name in setdiff(names(config), c(variables,"")))
+            data[[name]] <- rep(config[[name]], length.out=n)
+    }
     
     report(OL$Info, "Scheduling #{n} jobs")
     
@@ -56,9 +75,10 @@ ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, 
     
     buildArgs <- function (i)
     {
-        currentFlags <- ore.subst("(?<!\\\\)\\%([A-Za-z]+)", function(match) data[i,groups(match)], tractorFlags, all=TRUE)
+        index <- ifelse(repetitions > 0, 1, i)
+        currentFlags <- ore.subst("(?<!\\\\)\\%([A-Za-z]+)", function(match) data[index,groups(match)], tractorFlags, all=TRUE)
         currentFlags <- ore.subst("(?<!\\\\)\\%\\%", as.character(i), currentFlags, all=TRUE)
-        currentOptions <- ore.subst("(?<!\\\\)\\%([A-Za-z]+)", function(match) data[i,groups(match)], tractorOptions, all=TRUE)
+        currentOptions <- ore.subst("(?<!\\\\)\\%([A-Za-z]+)", function(match) data[index,groups(match)], tractorOptions, all=TRUE)
         currentOptions <- ore.subst("(?<!\\\\)\\%\\%", as.character(i), currentOptions, all=TRUE)
         return (es("#{currentFlags} #{scriptName} #{currentOptions}"))
     }
@@ -75,15 +95,20 @@ ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, 
         writeLines(args, argsFile)
         
         configPrefix <- file.path(tempDir, "config")
-        for (i in seq_len(n))
-            writeYaml(as.list(data[i,,drop=FALSE]), es("#{configPrefix}.#{i}.yaml"))
+        if (repetitions == 0)
+        {
+            for (i in seq_len(n))
+                writeYaml(as.list(data[i,,drop=FALSE]), es("#{configPrefix}.#{i}.yaml"))
+        }
         
         qsubScriptFile <- file.path(tempDir, "script")
         qsubScript <- c("#!/bin/sh",
                         "#$ -S /bin/bash",
                         es("TRACTOR_PLOUGH_ID=${SGE_TASK_ID}"),
                         es("TRACTOR_ARGS=`sed \"${SGE_TASK_ID}q;d\" #{argsFile}`"),
-                        es("#{tractorPath} -c #{configPrefix}.${SGE_TASK_ID}.yaml ${TRACTOR_ARGS}"))
+                        ifelse(repetitions > 0,
+                            es("#{tractorPath} -c #{configFile} ${TRACTOR_ARGS}"),
+                            es("#{tractorPath} -c #{configPrefix}.${SGE_TASK_ID}.yaml ${TRACTOR_ARGS}")))
         writeLines(qsubScript, qsubScriptFile)
         execute("chmod", es("+x #{qsubScriptFile}"))
         
@@ -97,9 +122,14 @@ ploughExperiment <- function (scriptName, configFiles, variables, tractorFlags, 
     else
     {
         parallelApply(seq_len(n), function(i) {
-            currentFile <- threadSafeTempFile()
-            writeYaml(as.list(data[i,,drop=FALSE]), currentFile)
-            on.exit(unlink(currentFile))
+            if (repetitions > 0)
+                currentFile <- configFile
+            else
+            {
+                currentFile <- threadSafeTempFile()
+                writeYaml(as.list(data[i,,drop=FALSE]), currentFile)
+                on.exit(unlink(currentFile))
+            }
             
             # FIXME: Using callExperiment() would be preferable, to allow interrupts, but flags will need to be parsed
             # callExperiment(exptName, args = NULL, configFiles = NULL, outputLevel = getOutputLevel(), ...)
